+++ solver_test.c
@@ -4,7 +4,32 @@
 #include <unistd.h>
 #include "mmio.h"
 
-static int matrixRead(matrixCOO* localMatrixCOO, float** xCompare_in, float** y_in, FILE *f)
+//static void matrix_vectorTest(matrixCOO *matrix, const double* vector_in, double* vector_out, 
+//		int testpoint){
+//	FILE *tFile;
+//	if(testpoint >= 0) {
+//		if ((tFile = fopen("testResult", "w")) == NULL){ 
+//			printf("file open error\n");
+//			exit(1);
+//		}
+//	}
+//	for(unsigned int i = 0; i < matrix->dimension; ++i) vector_out[i] = 0;
+//	for(unsigned int i = 0; i < matrix->totalNum; ++i){
+//		vector_out[matrix->I[i]] += matrix->V[i]*vector_in[matrix->J[i]];	
+//		if(matrix->I[i] == testpoint && testpoint >= 0){ fprintf(tFile, 
+//				"row is %d V is %f vec is %f accum is %f\n", 
+//				testpoint, matrix->V[i], vector_in[matrix->J[i]], vector_out[matrix->I[i]]);
+//		}
+//		//if(vector_out[matrix->I[i]] > 0 && vector_out[matrix->I[i]] < 1)
+//		//	printf("strange vector out\n");
+//
+//	}
+//	if(testpoint >= 0)
+//		fclose(tFile);
+//}
+
+
+static int matrixRead(matrixCOO* localMatrixCOO, double** xCompare_in, double** y_in, FILE *f)
 {
 	int _dimension, _N, _lowerNum;
 	int ret_code, totalNum, lowerNum;
@@ -14,13 +39,13 @@ static int matrixRead(matrixCOO* localMatrixCOO, float** xCompare_in, float** y_
 	lowerNum = _lowerNum;
 	totalNum = lowerNum*2-_dimension;
 	
-	*xCompare_in = (float*)malloc(_dimension*sizeof(float));
-	*y_in = (float*)malloc(_dimension*sizeof(float));
-	float*	xCompare = *xCompare_in;
-	float*	y = *y_in;
+	*xCompare_in = (double*)malloc(_dimension*sizeof(double));
+	*y_in = (double*)malloc(_dimension*sizeof(double));
+	double*	xCompare = *xCompare_in;
+	double*	y = *y_in;
 	/*The overall number of nozeros in this matrix*/
 	localMatrixCOO->totalNum = totalNum;
-	localMatrixCOO->nParts = ceil(((float) _dimension)/vectorCacheSize);
+	localMatrixCOO->nParts = ceil(((double) _dimension)/vectorCacheSize);
 	printf("parts is %d\n", localMatrixCOO->nParts);
 	localMatrixCOO->partBoundary = (int* )calloc(_dimension, sizeof(int));
 	localMatrixCOO->reorderList = (int* )calloc(_dimension, sizeof(int));
@@ -29,22 +54,22 @@ static int matrixRead(matrixCOO* localMatrixCOO, float** xCompare_in, float** y_
 
 	int* lowerI=(int *) malloc(lowerNum*sizeof(int));
 	int* lowerJ=(int *) malloc(lowerNum*sizeof(int));
-	float* lowerV=(float *) malloc(lowerNum*sizeof(float));
+	double* lowerV=(double *) malloc(lowerNum*sizeof(double));
 
 	localMatrixCOO->I=(int *) malloc(totalNum*sizeof(int));
 	localMatrixCOO->J=(int *) malloc(totalNum*sizeof(int));
-	localMatrixCOO->V=(float *) malloc(totalNum*sizeof(float));
-	localMatrixCOO->diag = (float *) malloc(_dimension*sizeof(float));
+	localMatrixCOO->V=(double *) malloc(totalNum*sizeof(double));
+	localMatrixCOO->diag = (double *) malloc(_dimension*sizeof(double));
 	int* numInRow = localMatrixCOO->numInRow;
 	int* I = localMatrixCOO->I;
 	int* J = localMatrixCOO->J;
-	float* V = localMatrixCOO->V;
-	float* diag = localMatrixCOO->diag;
+	double* V = localMatrixCOO->V;
+	double* diag = localMatrixCOO->diag;
 	
 	int tempI, tempJ;
-	float tempV;
+	double tempV;
 	for (int i=0; i<lowerNum; i++){
-		fscanf(f, "%d %d %f\n", &tempI, &tempJ, &tempV);
+		fscanf(f, "%d %d %lg\n", &tempI, &tempJ, &tempV);
 		lowerJ[i]=tempJ-1;  /* adjust from 1-based to 0-based */
 		lowerI[i]=tempI-1;
 		lowerV[i]=tempV;
@@ -116,9 +141,9 @@ int main(int argc, char* argv[])
 	MM_typecode matcode;
 	int MAXIter = 0;
 	FILE *f;
-	float *x;
-	float *y;
-	float *xCompare;
+	double *x;
+	double *y;
+	double *xCompare;
 	char fileName[100];
 	fileName[0] = '\0';
 	int oc;
@@ -194,18 +219,18 @@ int main(int argc, char* argv[])
 	matrixCOO localMatrixCOO;
 	matrixRead(&localMatrixCOO, &xCompare, &y, f);
 	fclose(f);
-	x = (float *) calloc(localMatrixCOO.dimension, sizeof(float));
+	x = (double *) calloc(localMatrixCOO.dimension, sizeof(double));
 
 	spmvHYB(&localMatrixCOO, y, x, MAXIter);
 	for (int i=0;i<10;i++)
 	{
 		printf("at %d x is %f x_compare is  %f\n",i + 30000, x[i + 30000], xCompare[i + 30000]);
 	}
-	memset(x, 0, sizeof(float)*localMatrixCOO.dimension);
+	memset(x, 0, sizeof(double)*localMatrixCOO.dimension);
 	return 0;
 
-	float *xReorder = (float* )calloc(localMatrixCOO.dimension, sizeof(float)); 
-	float *yReorder = (float* )calloc(localMatrixCOO.dimension, sizeof(float)); 
+	double *xReorder = (double* )calloc(localMatrixCOO.dimension, sizeof(double)); 
+	double *yReorder = (double* )calloc(localMatrixCOO.dimension, sizeof(double)); 
 	matrixReorder(&localMatrixCOO);
 	vectorReorder(localMatrixCOO.dimension, y, yReorder, localMatrixCOO.reorderList);
 
@@ -221,6 +246,9 @@ int main(int argc, char* argv[])
 	{
 		printf("at %d x is %f x_compare is  %f\n",i + 30000, x[i + 30000], xCompare[i + 30000]);
 	}
+	//check the correctness
+	//matrix_vectorTest(localMatrixCOO, y, x_2);
+
 	free(localMatrixCOO.I);
 	free(localMatrixCOO.J);
 	free(localMatrixCOO.V);
@@ -235,7 +263,7 @@ int main(int argc, char* argv[])
 	//interval2=(end_time2-start_time2)*1000/CLOCKS_PER_SEC;
 
 	//printf("time consuming CPU is %f, time consuming GPU is %f, speedup is %f\n", interval1, interval2, interval1/interval2);
-	//float Gflop=(totalNum*4+12*dimension)/interval1*1000*MAXIter;
+	//double Gflop=(totalNum*4+12*dimension)/interval1*1000*MAXIter;
 	//printf("error is %f, total num is %d, time is %f ms, Gflops is %f, final error is %f\n",result_error/dimension, totalNum, interval1, Gflop, error_track[MAXIter-1]*1000);
 	return 0;
 }
