+++ spmv.h
@@ -24,8 +24,8 @@ typedef struct _matrixCOO{
     int* numInRow2;//"real" numInRow for vetex goto blockELL
     int* I;
     int* J;
-    float* V;
-    float* diag;
+    double* V;
+    double* diag;
 	int* partBoundary;
 	int* reorderList;
 }matrixCOO;
@@ -39,13 +39,13 @@ typedef struct _matrixEHYB{
 	int* widthVecBlockELL;
 	int* biasVecBlockELL;
 	int16_t* colBlockELL;
-	float* valBlockELL;
+	double* valBlockELL;
 	int* partBoundary;
 	int* widthVecER;
 	int* rowVecER;
 	int* biasVecER; 
 	int* colER;
-	float* valER;
+	double* valER;
 }matrixEHYB;
 
 inline void init_cb(cb_s* in_s)
@@ -60,14 +60,14 @@ inline void init_cb(cb_s* in_s)
 
 extern "C"
 void spmvGPuEHYB(matrixCOO* localMatrix, 
-		const float* vectorIn, float* vectorOut,  
+		const double* vectorIn, double* vectorOut,  
 		const int MAXIter, int* realIter);
 
 void solverGPuUnprecondCUSPARSE(matrixCOO* localMatrix, 
-		const float *vector_in, float *vector_out,  
+		const double *vector_in, double *vector_out,  
 		const int MAXIter);
 
 void spmvHYB(matrixCOO* localMatrix, 
-		const float *vector_in, float *vector_out,  
+		const double *vector_in, double *vector_out,  
 		const int MAXIter);
 #endif
